+++ Crystal/CrystalCommand/BoundarySettingCommand.cpp
@@ -35,13 +35,13 @@ void BoundarySettingCommand::displayBoundarySetting(System::Windows::Forms::Data
 
 void BoundarySettingCommand::saveBoundarySetting(System::Windows::Forms::DataGridView^ view)
 {
-	const double minX = Convert::ToDouble( view->Rows[0]->Cells[0]->Value );
-	const double minY = Convert::ToDouble( view->Rows[0]->Cells[1]->Value );
-	const double minZ = Convert::ToDouble( view->Rows[0]->Cells[2]->Value );
+	const float minX = Convert::ToSingle( view->Rows[0]->Cells[0]->Value );
+	const float minY = Convert::ToSingle( view->Rows[0]->Cells[1]->Value );
+	const float minZ = Convert::ToSingle( view->Rows[0]->Cells[2]->Value );
 
-	const double maxX = Convert::ToDouble( view->Rows[1]->Cells[0]->Value );
-	const double maxY = Convert::ToDouble( view->Rows[1]->Cells[1]->Value );
-	const double maxZ = Convert::ToDouble( view->Rows[1]->Cells[2]->Value );
+	const float maxX = Convert::ToSingle( view->Rows[1]->Cells[0]->Value );
+	const float maxY = Convert::ToSingle( view->Rows[1]->Cells[1]->Value );
+	const float maxZ = Convert::ToSingle( view->Rows[1]->Cells[2]->Value );
 
 	 ApplicationSettings::get()->simulationSetting->boundaryBox = Box( Crystal::Geom::Vector3d( minX, minY, minZ), 
 		Crystal::Geom::Vector3d( maxX, maxY, maxZ) );
