+++ Crystal/CrystalPhysics/Fluid.cpp
@@ -11,7 +11,7 @@ using namespace Crystal::Geom;
 using namespace Crystal::Physics;
 
 
-Fluid::Fluid(const int id, const double density, const double gasConstant, const double viscosityCoefficient, ParticleFactory* particleFactory) :
+Fluid::Fluid(const int id, const float density, const float gasConstant, const float viscosityCoefficient, ParticleFactory* particleFactory) :
 PhysicsObject( id, density, gasConstant, viscosityCoefficient,particleFactory )
 {
 }
@@ -20,7 +20,7 @@ Fluid::~Fluid(void)
 {
 }
 
-void Fluid::integrateTime(const double proceedTime)
+void Fluid::integrateTime(const float proceedTime)
 {
 	const ParticleVector& particles = getParticles();
 	BOOST_FOREACH( Particle* particle, particles ) {
