+++ Crystal/CrystalPhysics/Obstacle.h
@@ -11,15 +11,15 @@ class Obstacle : public PhysicsObject
 {
 public:
 	
-	Obstacle(const int id, const double density, const double gasConstant, const double viscosityCoefficient, ParticleFactory* particleFactory);
+	Obstacle(const int id, const float density, const float gasConstant, const float viscosityCoefficient, ParticleFactory* particleFactory);
 
 	~Obstacle();
 
 	virtual Type getType() { return PhysicsObject::Obstacle; }
 
-	virtual void integrateTime(const double proceedTime){};
+	virtual void integrateTime(const float proceedTime){};
 
-	virtual void enforce(const double proceedTime){};
+	virtual void enforce(const float proceedTime){};
 };
 
 	}
