+++ Crystal/CrystalPhysics/SearchParticle.h
@@ -25,13 +25,13 @@ public:
 	{
 	}
 
-	SearchParticle(Particle* particle, const double effectLength): particle(particle){
+	SearchParticle(Particle* particle, const float effectLength): particle(particle){
 		setGridID( particle->center, effectLength );
 	}
 
 	~SearchParticle(void){};
 
-	void setGridID(const Geom::Vector3d& point, const double effectLength) {
+	void setGridID(const Geom::Vector3d& point, const float effectLength) {
 		int gridX = static_cast<int>( point.getX() / effectLength );
 		int gridY = static_cast<int>( point.getY() / effectLength );
 		int gridZ = static_cast<int>( point.getZ() / effectLength );
